ToTW: Automate Software Workflows with GitHub Actions (#16)
This PR adds a new tip of the week blog post titled "Automate Software
Workflows with GitHub Actions". The focus here was on providing a
baseline understanding and inspiration for software workflow engines,
Github Actions, and environments reproducibility (with context to
software workflows). As part of this post, supplemental demonstration
content was created via CU-DBMI/demo-github-actions#1 to help show how
features look in real-life and reduce the barriers to entry for learning
about this feature from Github alongside third-party tools.

Thank you in advance for your thoughts, comments, and suggestions!

---------

Co-authored-by: Faisal Alquaddoomi <falquaddoomi@gmail.com>
Co-authored-by: Vincent Rubinetti <8326331+vincerubinetti@users.noreply.github.com>