Packaging as publishing blog post (#22)
This PR adds content surrounding a new blog post titled: "Tip of the
Week: Python Packaging as Publishing". My hope here is to cover the
"justification" and "preparation" portions of packaging, covering
"distribution" (i.e. PyPI, etc) in a separate post in the future.

Thanks in advance for any feedback and thoughts you may have on this!